Add files via upload
Here is a form that requires a name, an email and a image. The image is filtered to check the width and height and resize it according to the requirements. Also the form data is saved to the database and then it is shown on another page. The project was made without any framework, only plain Js, Css, Html, Php and MySql